- Añade proteccion de rutas a los endpoints
- Nuevo claim JWT lectorID opcional para evitar queries extras a la DB
- Nuevo rol auxiliar con permisos de lectura y acciones de circulacion (renovacion de prestamos, altas, etc)

+++ routes/articulo.php
@@ -14,30 +14,30 @@ $articuloRepository = new ArticuloRepository();
 $articuloService = new ArticuloService($articuloRepository);
 $articuloController = new ArticuloController($articuloService);
 
-$router->get('/articulos', function () use ($articuloController) {
+$router->get('/articulos', withRole(['admin', 'auxiliar', 'lector'], function () use ($articuloController) {
     $articuloController->getAll();
-});
+}));
 
-$router->get('/articulos/{idArticulo}/temas', function ($idArticulo) use ($articuloController) {
+$router->get('/articulos/{idArticulo}/temas', withRole(['admin', 'lector'], function ($idArticulo) use ($articuloController) {
     $articuloController->getTemaTitlesByArticulo($idArticulo);
-});
+}));
 
-$router->post('/articulos/{idArticulo}/temas/{idTema}', function ($idArticulo, $idTema) use ($articuloController) {
+$router->post('/articulos/{idArticulo}/temas/{idTema}', withRole(['admin'], function ($idArticulo, $idTema) use ($articuloController) {
     $articuloController->addTemaToArticulo($idArticulo, $idTema);
-});
+}));
 
-$router->delete('/articulos/{idArticulo}/temas/{idTema}', function ($idArticulo, $idTema) use ($articuloController) {
+$router->delete('/articulos/{idArticulo}/temas/{idTema}', withRole(['admin'], function ($idArticulo, $idTema) use ($articuloController) {
     $articuloController->deleteTemaFromArticulo($idArticulo, $idTema);
-});
+}));
 
-$router->get('/articulos/{id}', function ($id) use ($articuloController) {
+$router->get('/articulos/{id}', withRole(['admin', 'auxiliar', 'lector'], function ($id) use ($articuloController) {
     $articuloController->getById($id);
-});
+}));
 
-$router->patch('/articulos/{id}', function ($id) use ($articuloController) {
+$router->patch('/articulos/{id}', withRole(['admin'], function ($id) use ($articuloController) {
     $articuloController->patchArticulo($id);
-});
+}));
 
-$router->delete('/articulos/{id}', function ($id) use ($articuloController) {
+$router->delete('/articulos/{id}', withRole(['admin'], function ($id) use ($articuloController) {
     $articuloController->deleteArticulo($id);
-});
+}));
